Crate rustros_tf

This is a rust port of the ROS tf library. It is intended for being used in robots to help keep track of multiple coordinate frames and is part of a larger suite of rust libraries that provide support for various robotics related functionality.

Example usage:

fn main() {
    rosrust::init("listener");
    let listener = TfListener::new();
    
    let rate = rosrust::rate(1.0);
    while rosrust::is_ok() {
        let tf = listener.lookup_transform("camera", "base_link", ros::Time::now());
        println!("{:?}", tf);
        rate.sleep();
    }
}

Modules

msg

Structs

TfListener
This struct tries to be the same as the C++ version of TransformListener. Use this struct to lookup transforms.

Enums

TfError
Enumerates the different types of errors

Functions

get_inverse
Calculates the inverse of a ros transform